requirements.txt
apispec==2.0.2
apispec-oneofschema==2.1.0
apispec-webframeworks==0.4.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.0
biocommons.seqrepo==0.5.1
biopython==1.69
bioutils==0.5.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
coloredlogs==10.0
configparser==4.0.2
cssselect==1.0.3
decorator==4.4.1
enum34==1.1.6
fake-useragent==0.1.11
Flask==1.1.1
Flask-Cors==3.0.8
hgvs==1.3.0.post0
humanfriendly==4.18
idna==2.8
importlib-metadata==0.23
iniherit==0.3.9
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
lazy-object-proxy==1.4.2
lxml==4.3.4
MarkupSafe==1.1.1
marshmallow==3.2.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==2.0.1
mccabe==0.6.1
more-itertools==7.2.0
mypy==0.761
mypy-extensions==0.4.3
numpy==1.17.2
packaging==19.2
parse==1.12.0
Parsley==1.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.13.0
prompt-toolkit==1.0.18
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.0
pyee==6.0.0
Pygments==2.4.2
pylint==2.3.1
pyparsing==2.4.5
pyppeteer==0.0.25
pyquery==1.4.0
pysam==0.15.2
pytest==5.2.2
PyYAML==5.1.2
requests==2.22.0
requests-html==0.10.0
seqrepo==0.0.0
simplegeneric==0.8.1
six==1.12.0
soupsieve==1.9.2
text-unidecode==1.2
tqdm==4.32.2
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.3
uWSGI==2.0.18
w3lib==1.20.0
wcwidth==0.1.7
websockets==8.0.1
Werkzeug==0.16.0
wrapt==1.11.2
yoyo-migrations==6.1.0
zipp==0.6.0
